fixed modal and pattern
